+++ NEWS.md
@@ -1,3 +1,12 @@
+# Version 0.1.1 [current]
+
+### Minor changes
+* Add `transform.attr` parameter to user control transformation of datatypes
+* Measures can return NULL in fails or NA if they are not possible computing
+
+### Bugfixes
+* Support minority class with an unique instance
+
 # First release of **mfe**:
 
 * General meta-features